:truck: #1 Rename(parser): parser_print 로 테스트 코드 분리
